refactor(playground): split chat-message into per-part components
The chat-message module had grown to ~770 lines housing the role
dispatcher, every part renderer (reasoning, tool call, tool output,
step start, joined assistant text), and several components
(mcp-app-view, attachment-preview, image-modal, token-usage,
no-content-message) that only render inside chat messages. Each piece
moves into a focused sibling under components/chat-message/, with the
public entry kept at components/chat-message/index.tsx so existing
`import { ChatMessage } from "./chat-message"` callsites resolve
unchanged. The folder/index layout matches the existing card-mcp-server,
form-run-from-registry, and groups-manager idiom.

The dispatcher itself shrinks to a small `memo(ChatMessageImpl)` that
hands off to UserMessage / AssistantMessage, so streaming ticks only
re-render the row whose message ref actually changed and sibling rows
skip via shallow prop equality.
