+++ renderer/src/features/chat/components/chat-message/__tests__/mcp-app-view.test.tsx
