+++ renderer/src/features/chat/components/chat-message/attachment-preview.tsx
