+++ renderer/src/features/chat/components/chat-message/image-modal.tsx
