+++ renderer/src/features/chat/components/chat-message/mcp-app-view.tsx
